Phone: Fix race condition in BOOT_COMPLETED intent
BOOT_COMPLETED was triggering a read (and write) of SIM extra
fields, regardless of whether the SIM was actually ready for it
or not.
If the boot happened to complete before the radio layer
was ready, we were trying to do iccIO before we were allowed to.
This resulted in errors for most phones, but on at least one it caused
the whole iccIO interface to lock up until the next reboot (where the
same would probably happen again)

So... trigger the write with the SIM_READY intent itself instead of
using BOOT_COMPLETED

Change-Id: I027fab6000cae0ed94ea1513553d883fde99ca75